Tratando o erro de entidade nao encontrada
