+++ src/main/java/med/voll/api/domain/medico/DadosCadastroMedicos.java
@@ -1,24 +1,28 @@
-package med.voll.api.paciente;
+package med.voll.api.domain.medico;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import med.voll.api.endereco.DadosEndereco;
+import med.voll.api.domain.endereco.DadosEndereco;
 
-public record DadosCadastroPacientes(
+public record DadosCadastroMedicos(
+
         @NotBlank
         String nome,
         @NotBlank
         @Email
         String email,
+
         @NotBlank
-        @Pattern(regexp = "\\d{9}")
         String telefone,
         @NotBlank
-        @Pattern(regexp = "\\d{6}")
-        String cpf,
+        @Pattern(regexp = "\\d{4,6}")
+        String crm,
+        @NotNull
+        Especialidade especialidade,
         @NotNull @Valid
-        DadosEndereco endereco) {
+        DadosEndereco endereco
+    ) {
 }
