+++ src/main/java/med/voll/api/domain/medico/Especialidade.java
@@ -1,4 +1,4 @@
-package med.voll.api.medico;
+package med.voll.api.domain.medico;
 
 public enum Especialidade {
     ORTOPEDIA,
