+++ src/main/java/med/voll/api/domain/medico/MedicoRepository.java
@@ -1,11 +1,9 @@
-package med.voll.api.medico;
+package med.voll.api.domain.medico;
 
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 
 public interface MedicoRepository extends JpaRepository<Medico, Long> {
