+++ src/main/java/med/voll/api/domain/paciente/DadosCadastroPacientes.java
@@ -1,28 +1,24 @@
-package med.voll.api.medico;
+package med.voll.api.domain.paciente;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import med.voll.api.endereco.DadosEndereco;
+import med.voll.api.domain.endereco.DadosEndereco;
 
-public record DadosCadastroMedicos(
-
+public record DadosCadastroPacientes(
         @NotBlank
         String nome,
         @NotBlank
         @Email
         String email,
-
         @NotBlank
+        @Pattern(regexp = "\\d{9}")
         String telefone,
         @NotBlank
-        @Pattern(regexp = "\\d{4,6}")
-        String crm,
-        @NotNull
-        Especialidade especialidade,
+        @Pattern(regexp = "\\d{6}")
+        String cpf,
         @NotNull @Valid
-        DadosEndereco endereco
-    ) {
+        DadosEndereco endereco) {
 }
