+++ src/main/java/med/voll/api/domain/paciente/DadosListagemPaciente.java
@@ -1,6 +1,4 @@
-package med.voll.api.paciente;
-
-import med.voll.api.medico.DadosListagemMedico;
+package med.voll.api.domain.paciente;
 
 public record DadosListagemPaciente(Long id, String nome, String email, String cpf) {
