+++ src/main/java/med/voll/api/domain/paciente/PacienteRepository.java
@@ -1,10 +1,8 @@
-package med.voll.api.paciente;
+package med.voll.api.domain.paciente;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface PacienteRepository extends JpaRepository<Paciente, Long> {
     Page<Paciente> findByAtivoTrue(Pageable pageable);
