+++ src/main/java/med/voll/api/infra/TratadorDeErros.java
@@ -1,0 +1,16 @@
+package med.voll.api.infra;
+
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class TratadorDeErros {
+
+    @ExceptionHandler(EntityNotFoundException.class)
+    public ResponseEntity tratarErro404(){
+        return ResponseEntity.notFound().build();
+    }
+
+}
